change default max file size and remove in mem clone
The max file size has been dropped to 10MB from to give a more realistic use case. 

The flags for in-mem-clone have bencommented out, though the default remains. This is due to the functionality not working with the way we are using ignore path and file extensions. There is a story around adding this back in but it reques some thinking about how we handle the filesystem and access it.