Develop (#97)
* clean up flags

* update

* update

* update flag strings

* formating issues

* webserver enable/disable flaf

* enhanced debug and header output

* csv and json output

* bring in ability to update signatures to a specific version

* Update sample configuration  and documentation (#91)

Update the yaml configuration file example, as well as the description in step 1 and the yaml example in the README.md file.

According to implementations found in scanLocalGitRepo.go:69 and scanLocalPath.go:85, the sample.yml and the quickstart instructions weren't referring to the updated parameter names.

* cleanup

* fix the progress bar

Still not quite right due to threads but better

* remove dead functions

* Updated the changelog

* implement persistent flags

move the flags that are common to all the commands into the root command. The updateRules command does not adhere to all the flags but for know we will deal with it.

Signed-off-by: Matty <urlugal@gmail.com>

* Cleanup

Signed-off-by: Matty <urlugal@gmail.com>

* implement persistent flags (#95)

move the flags that are common to all the commands into the root command. The updateRules command does not adhere to all the flags but for know we will deal with it.

Signed-off-by: Matty <urlugal@gmail.com>

Co-authored-by: Santiago Botta <sgobotta@gmail.com>
Co-authored-by: Matty <matty.j@innocentlivesfoundation.org>